Declared all dependencies in Version Catalog

+++ app/src/main/java/com/victor/calendar/ui/home/HomeScreen.kt
@@ -103,7 +103,7 @@ fun calculateTopPadding(
     dayStart: Long,
     eventStart: Long
 ): Float {
-    return (((eventStart - dayStart) / MILLIS_IN_DAY.toFloat()) * 1920)
+    return ((((eventStart - dayStart).toFloat()) / MILLIS_IN_DAY.toFloat()) * 1920)
 }
 
 fun calculateHourHeight(
